Initial implementation of futures-based batch verification.
This commit incidentally includes an optimization for batch verification that
improves performance when verifying multiple signatures from the same public
key.

I'm not totally happy with a few things about this API, however.  Currently,
the actual batch computation is performed only when the inherent `finalize()`
method is called on the service.  However, this doesn't work well with `tower`
layering, because once the service is wrapped, the inherent method is no longer
available.  Another option would be for the batching service to be created with
a batch size parameter, automatically resolving the batch computation whenever
the batch size was reached.  This improves latency but does not solve the
problem of finalizing the batch, since unless there's guaranteed to be a
constant stream of verification requests, a tail of requests may be left
unresolved.  A third option would be for the service to return some kind of
control handle on creation that would allow shutdown via a channel or
something, but this is also unsatisfying because it means that the service has
to be listening for a shutdown signal.  A fourth option would be to have a
batch size parameter but customize the `Drop` impl to finalize all pending
requests on drop; this would potentially allow "flushing" pending requests by
manually dropping the service in a way that would still be possible when using
tower wrappers.